+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- #3601, Fix jwt error return status `400` for invalid role - @taimoorzaeem
+- Fix jwt error returning HTTP status `400` for invalid role by @taimoorzaeem in #3601
 
 ## [13.0.0] - 2025-05-08